+++ src/main/java/br/jus/jfsp/nuit/contadoria/models/IndicesSalarios.java
@@ -19,6 +19,7 @@ import javax.persistence.Id;
 import javax.persistence.Index;
 import javax.persistence.PrePersist;
 import javax.persistence.Table;
+import java.math.BigDecimal;
 import java.util.Calendar;
 
 @Entity
@@ -46,7 +47,7 @@ public class IndicesSalarios extends BaseIndice {
 	@GeneratedValue(strategy = GenerationType.IDENTITY, generator = "gen_indices_atrasados")
 	private Long id;
 
-	public IndicesSalarios(Double indice, String descricao, Calendar data) {
+	public IndicesSalarios(BigDecimal indice, String descricao, Calendar data) {
 		super();
 		this.data = data;
 		this.descricao = descricao;
